chore: Spring과 관련없는 코드 삭제 및 JDBC 테스트 코드 추가
