+++ src/main/java/chess/ChessApplication.java
@@ -4,9 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringChessApplication {
+public class ChessApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(SpringChessApplication.class, args);
+        SpringApplication.run(ChessApplication.class, args);
     }
 }
