+++ src/main/java/chess/dao/GameDaoImpl.java
@@ -1,79 +1,56 @@
 package chess.dao;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Repository;
+
 import chess.dto.response.ChessGameDto;
-import chess.exception.DatabaseException;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 
-public class GameDaoImpl extends Dao implements GameDao {
-
+@Repository
+public class GameDaoImpl implements GameDao {
     private static final String TABLE_NAME = "game";
     private static final String WHITE_TURN = "WHITE";
     private static final String BLACK_TURN = "BLACK";
 
+    private JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    public GameDaoImpl(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
     public ChessGameDto getGame(String gameId) {
         String query = String.format("SELECT turn FROM %s WHERE id = ?", TABLE_NAME);
-
-        try (Connection connection = getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setString(1, gameId);
-            ResultSet resultSet = preparedStatement.executeQuery();
-
-            boolean hasNext = resultSet.next();
-            if (!hasNext) {
-                throw new IllegalArgumentException("해당하는 게임 ID가 존재하지 않습니다.");
-            }
-
-            return ChessGameDto.of(gameId, resultSet.getString("turn"));
-        } catch (SQLException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        String turn = jdbcTemplate.queryForObject(query, (resultSet, rowNum) -> resultSet.getString("turn"), gameId);
+        return ChessGameDto.of(gameId, turn);
     }
 
+    @Override
     public void createGame(String gameId) {
         String query = String.format("INSERT INTO %s VALUES (?, 'WHITE')", TABLE_NAME);
-
-        try (Connection connection = getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setString(1, gameId);
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        jdbcTemplate.update(query, gameId);
     }
 
+    @Override
     public void deleteGame(String gameId) {
         String query = String.format("DELETE FROM %s WHERE id = ?", TABLE_NAME);
-
-        try (Connection connection = getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setString(1, gameId);
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        jdbcTemplate.update(query, gameId);
     }
 
+    @Override
     public void updateTurnToWhite(String gameId) {
         updateTurn(gameId, WHITE_TURN);
     }
 
+    @Override
     public void updateTurnToBlack(String gameId) {
         updateTurn(gameId, BLACK_TURN);
     }
 
     private void updateTurn(String gameId, String turn) {
         String query = String.format("UPDATE %s SET turn = ? WHERE id = ?", TABLE_NAME);
-
-        try (Connection connection = getConnection()) {
-            PreparedStatement preparedStatement = connection.prepareStatement(query);
-            preparedStatement.setString(1, turn);
-            preparedStatement.setString(2, gameId);
-            preparedStatement.executeUpdate();
-        } catch (SQLException e) {
-            throw new DatabaseException(e.getMessage());
-        }
+        jdbcTemplate.update(query, turn, gameId);
     }
+
 }
