--- src/main/java/chess/exception/DatabaseException.java
@@ -1,7 +1,0 @@
-package chess.exception;
-
-public class DatabaseException extends RuntimeException {
-    public DatabaseException(String message) {
-        super(message);
-    }
-}
