+++ src/test/java/chess/dao/BoardDaoImplTest.java
@@ -1,6 +1,13 @@
 package chess.dao;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
 
 import chess.domain.piece.Piece;
 import chess.domain.piece.PieceColor;
@@ -12,11 +19,8 @@ import chess.dto.request.CreatePieceDto;
 import chess.dto.request.DeletePieceDto;
 import chess.dto.request.UpdatePiecePositionDto;
 import chess.dto.response.BoardDto;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 
+@JdbcTest
 class BoardDaoImplTest {
     private static final String GAME_ID = "test-game";
     private static final XAxis X_AXIS = XAxis.A;
@@ -28,11 +32,31 @@ class BoardDaoImplTest {
 
     private BoardDaoImpl boardDao;
 
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
+
     @BeforeEach
     void setUp() {
-        boardDao = new BoardDaoImpl();
+        boardDao = new BoardDaoImpl(jdbcTemplate);
 
-        GameDaoImpl gameDao = new GameDaoImpl();
+        GameDaoImpl gameDao = new GameDaoImpl(jdbcTemplate);
+
+        jdbcTemplate.execute("DROP TABLE game, board IF EXISTS");
+        jdbcTemplate.execute("CREATE TABLE game("
+            + "id   VARCHAR(36) NOT NULL,"
+            + "turn ENUM('WHITE', 'BLACK'),"
+            + "PRIMARY KEY (id))"
+        );
+
+        jdbcTemplate.execute("CREATE TABLE board("
+            + "game_id     VARCHAR(36) NOT NULL,"
+            + "x_axis      ENUM('1', '2', '3', '4', '5', '6', '7', '8'),"
+            + "y_axis      ENUM('1', '2', '3', '4', '5', '6', '7', '8'),"
+            + "piece_type  ENUM('PAWN', 'ROOK', 'KNIGHT', 'BISHOP', 'QUEEN', 'KING'),"
+            + "piece_color ENUM('WHITE', 'BLACK'),"
+            + "PRIMARY KEY (game_id, x_axis, y_axis),"
+            + "FOREIGN KEY (game_id) REFERENCES game (id) ON DELETE CASCADE)"
+        );
         gameDao.createGame(GAME_ID);
     }
 
@@ -51,7 +75,7 @@ class BoardDaoImplTest {
     void createPiece() {
         // given
         CreatePieceDto createPieceDto = CreatePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS),
-                new Piece(PIECE_TYPE, PIECE_COLOR));
+            new Piece(PIECE_TYPE, PIECE_COLOR));
 
         // when & then
         boardDao.createPiece(createPieceDto);
@@ -63,7 +87,7 @@ class BoardDaoImplTest {
         // given
         DeletePieceDto deletePieceDto = DeletePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS));
         boardDao.createPiece(
-                CreatePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS), new Piece(PIECE_TYPE, PIECE_COLOR)));
+            CreatePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS), new Piece(PIECE_TYPE, PIECE_COLOR)));
 
         // when & then
         boardDao.deletePiece(deletePieceDto);
@@ -74,17 +98,11 @@ class BoardDaoImplTest {
     void updatePiecePosition() {
         // given
         UpdatePiecePositionDto updatePiecePositionDto = UpdatePiecePositionDto.of(GAME_ID, X_AXIS, Y_AXIS, X_AXIS_2,
-                Y_AXIS_2);
+            Y_AXIS_2);
         boardDao.createPiece(
-                CreatePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS), new Piece(PIECE_TYPE, PIECE_COLOR)));
+            CreatePieceDto.of(GAME_ID, Position.of(X_AXIS, Y_AXIS), new Piece(PIECE_TYPE, PIECE_COLOR)));
 
         // then
         boardDao.updatePiecePosition(updatePiecePositionDto);
-    }
-
-    @AfterEach
-    void tearDown() {
-        GameDaoImpl gameDao = new GameDaoImpl();
-        gameDao.deleteGame(GAME_ID);
     }
 }
